# Optimize `IOSource#read_until` method by using `StringScanner#check_until(string)`